refactored message class and changed its name to peep
